split svg-view to svg-editor and svg-selector
