+++ index.html
@@ -8,13 +8,11 @@
     <link rel="stylesheet" type="text/css" href="./index.css" />
     <title>Vite + Lit + TS</title>
     <script type="module" src="/src/file-access.ts"></script>
-    <script type="module" src="/src/svg-view.ts"></script>
-    <script type="module" src="/src/svg-scale-ruller.ts"></script>
+    <script type="module" src="/src/svg-editor.ts"></script>
   </head>
 
   <body>
-    <file-access type=".svg, image/svg+xml"><svg-view></svg-view></file-access>
-    <svg-scale-ruller></svg-scale-ruller>
+    <file-access type=".svg, image/svg+xml"><svg-editor></svg-editor></file-access>
   </body>
 
 </html>
